+++ src/IECoreDelight/GeometryAlgo.cpp
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2017, John Haddon. All rights reserved.
+//  Copyright (c) 2017, Image Engine Design Inc. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -32,14 +32,17 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "GafferDelight/IECoreDelightPreview/NodeAlgo.h"
-#include "GafferDelight/IECoreDelightPreview/ParameterList.h"
+#include "IECoreDelight/NodeAlgo.h"
+#include "IECoreDelight/ParameterList.h"
 
-#include "IECoreScene/MeshPrimitive.h"
+#include "GafferScene/Private/IECoreScenePreview/Geometry.h"
+
+#include "IECore/SimpleTypedData.h"
 
 #include <nsi.h>
 
 using namespace std;
+using namespace Imath;
 using namespace IECore;
 using namespace IECoreScene;
 using namespace IECoreDelight;
@@ -47,65 +50,26 @@ using namespace IECoreDelight;
 namespace
 {
 
-const char *g_catmullClark = "catmull-clark";
-
-void staticParameters( const IECoreScene::MeshPrimitive *mesh, ParameterList &parameters )
+bool convert( const IECoreScenePreview::Geometry *geometry, NSIContext_t context, const char *handle )
 {
-	parameters.add( "nvertices", mesh->verticesPerFace() );
-
-	if( mesh->interpolation() == "catmullClark" )
+	if( geometry->getType() != "dl:environment" )
 	{
-		parameters.add( {
-			"subdivision.scheme",
-			&g_catmullClark,
-			NSITypeString,
-			0,
-			1,
-			0
-		} );
+		return false;
 	}
-}
 
-bool convertStatic( const IECoreScene::MeshPrimitive *mesh, NSIContext_t context, const char *handle )
-{
-	NSICreate( context, handle, "mesh", 0, nullptr );
+	NSICreate( context, handle, "environment", 0, nullptr );
 
-	ParameterList parameters;
-	staticParameters( mesh, parameters );
-	NodeAlgo::primitiveVariableParameterList( mesh, parameters, mesh->vertexIds() );
-
-	NSISetAttribute( context, handle, parameters.size(), parameters.data() );
-
-	return true;
-}
-
-bool convertAnimated( const vector<const IECoreScene::MeshPrimitive *> &meshes, const vector<float> &times, NSIContext_t context, const char *handle )
-{
-	NSICreate( context, handle, "mesh", 0, nullptr );
-
-	ParameterList parameters;
-	staticParameters( meshes.front(), parameters );
-
-	vector<ParameterList> animatedParameters;
-	NodeAlgo::primitiveVariableParameterLists(
-		vector<const Primitive *>( meshes.begin(), meshes.end() ),
-		parameters, animatedParameters,
-		meshes.front()->vertexIds()
-	);
-
-	NSISetAttribute( context, handle, parameters.size(), parameters.data() );
-
-	if( !animatedParameters.empty() )
+	if( const FloatData *angleData = geometry->parameters()->member<const FloatData>( "angle" ) )
 	{
-		for( size_t i = 0, e = animatedParameters.size(); i < e; ++i )
-		{
-			NSISetAttributeAtTime( context, handle, times[i], animatedParameters[i].size(), animatedParameters[i].data() );
-		}
+		const double angle = angleData->readable();
+		ParameterList parameters;
+		parameters.add( { "angle", &angle, NSITypeDouble, 0, 1, 0 } );
+		NSISetAttribute( context, handle, parameters.size(), parameters.data() );
 	}
 
 	return true;
 }
 
-NodeAlgo::ConverterDescription<MeshPrimitive> g_description( convertStatic, convertAnimated );
+NodeAlgo::ConverterDescription<IECoreScenePreview::Geometry> g_description( convert );
 
 } // namespace
